modules/openstack/secgroups/*: Allow ingress traffic within cluster
Relax the security group rules to allow ingress on all ports on the
internal subnet used by the k8s cluster. This subnet is not exposed
to the internet, and network policy can be applied directly to the k8s
cluster via NetworkPolicy objects, so the security posture remains
essentially unchanged.